requirements.txt
aih-dev==0.2.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
PyJWT==1.7.1
requests==2.21.0
urllib3==1.24.3